Struct lottie2gif::Animation

pub struct Animation(_);

A lottie animation.

Implementations

impl Animation

pub fn from_file<P>(path: P) -> Option<Animation> where
P: AsRef<Path>,

Read a lottie animation from file. This file needs to be in JSON format; if you want to read telegram’s tgs files, you need to decompress them first.

pub fn from_data<P>(
data: String,
key: String,
resource_path: P
) -> Option<Animation> where
P: AsRef<Path>,

Read a file from memory. External resources are resolved relative to resource_path.

pub fn size(&self) -> Size

Return the default viewport size of this animation.

pub fn duration(&self) -> f64

Return the total duration of this animation in seconds.

pub fn totalframe(&self) -> u64

Return the total number of frames in this animation.

pub fn framerate(&self) -> f64

Return the default framerate of this animation.

pub fn render_tree(&mut self, frame_num: u64, size: Size) -> LayerNode

pub fn frame_at_pos(&self, pos: f32) -> u64

Maps position to frame number and returns it.

pub fn render(
&mut self,
frame_num: u64,
buffer: &mut Vec<Argb, Global>,
size: Size
) -> Result<(), RenderError>

Render the contents of a frame into the buffer at a certain viewport size.

The buffer’s capacity must be at least size.width * size.height. It’s initial length or content doesn’t matter. The first size.width * size.height bytes of the buffer will be written to; and it’s length will be set exactly to size.width * size.height.

This operation will fail only if the buffer’s capacity isn’t large enough.